big_data_assignment/requirements.txt
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.0
xgboost==2.0.3
pyyaml==6.0.1
joblib==1.3.2
duckdb==0.10.3
pyspark==3.5.0
pyarrow==15.0.0
notebook==7.0.7
ipykernel==6.29.0
jupyterlab==4.1.0
